add global shortcut handlers to routes logic
