+++ src/entries/popup/Routes.tsx
@@ -3,11 +3,14 @@ import * as React from 'react';
 import { matchRoutes, useLocation } from 'react-router-dom';
 
 import { i18n } from '~/core/languages';
+import { useCurrentAddressStore } from '~/core/state';
 import { POPUP_DIMENSIONS } from '~/core/utils/dimensions';
 import { Box } from '~/design-system';
 import { AnimatedRoute } from '~/design-system/components/AnimatedRoute/AnimatedRoute';
 
 import { FullScreenBackground } from './components/FullScreen/FullScreenBackground';
+import { useAccounts } from './hooks/useAccounts';
+import { useKeyboardShortcut } from './hooks/useKeyboardShortcut';
 import { CreatePassword } from './pages/createPassword';
 import { Home } from './pages/home';
 import { ConnectedApps } from './pages/home/ConnectedApps';
@@ -102,6 +105,7 @@ const ROUTE_DATA = [
       </AnimatedRoute>
     ),
     background: FullScreenBackground,
+    disableGlobalShortcuts: true,
   },
   {
     path: ROUTES.READY,
@@ -130,6 +134,7 @@ const ROUTE_DATA = [
       </AnimatedRoute>
     ),
     background: FullScreenBackground,
+    disableGlobalShortcuts: true,
   },
   {
     path: ROUTES.WATCH,
@@ -146,6 +151,7 @@ const ROUTE_DATA = [
       </AnimatedRoute>
     ),
     background: FullScreenBackground,
+    disableGlobalShortcuts: true,
   },
   {
     path: ROUTES.IMPORT,
@@ -162,6 +168,7 @@ const ROUTE_DATA = [
       </AnimatedRoute>
     ),
     background: FullScreenBackground,
+    disableGlobalShortcuts: true,
   },
   {
     path: ROUTES.IMPORT__SELECT,
@@ -175,6 +182,7 @@ const ROUTE_DATA = [
       </AnimatedRoute>
     ),
     background: FullScreenBackground,
+    disableGlobalShortcuts: true,
   },
   {
     path: ROUTES.IMPORT__EDIT,
@@ -189,6 +197,7 @@ const ROUTE_DATA = [
       </AnimatedRoute>
     ),
     background: FullScreenBackground,
+    disableGlobalShortcuts: true,
   },
   {
     path: ROUTES.UNLOCK,
@@ -198,6 +207,7 @@ const ROUTE_DATA = [
       </AnimatedRoute>
     ),
     background: FullScreenBackground,
+    disableGlobalShortcuts: true,
   },
   {
     path: ROUTES.SEED_BACKUP_PROMPT,
@@ -214,6 +224,7 @@ const ROUTE_DATA = [
       </AnimatedRoute>
     ),
     background: FullScreenBackground,
+    disableGlobalShortcuts: true,
   },
   {
     path: ROUTES.SEED_REVEAL,
@@ -230,6 +241,7 @@ const ROUTE_DATA = [
       </AnimatedRoute>
     ),
     background: FullScreenBackground,
+    disableGlobalShortcuts: true,
   },
   {
     path: ROUTES.SEED_VERIFY,
@@ -243,6 +255,7 @@ const ROUTE_DATA = [
       </AnimatedRoute>
     ),
     background: FullScreenBackground,
+    disableGlobalShortcuts: true,
   },
   {
     path: ROUTES.CREATE_PASSWORD,
@@ -259,6 +272,7 @@ const ROUTE_DATA = [
       </AnimatedRoute>
     ),
     background: FullScreenBackground,
+    disableGlobalShortcuts: true,
   },
   {
     path: ROUTES.QR_CODE,
@@ -636,7 +650,7 @@ const matchingRoute = (pathName: string) => {
 export function Routes() {
   const location = useLocation();
   React.useEffect(() => {
-    // need to wait a tick for the page to render=
+    // need to wait a tick for the page to render
     setTimeout(() => {
       window.scrollTo(0, 0);
     }, 0);
@@ -669,6 +683,9 @@ function CurrentRoute(props: { pathname: string }) {
   const previousMatch = matchingRoute(state?.from || '');
   const previousElement = previousMatch?.element;
   const previousDirection = previousElement?.props.direction;
+
+  useGlobalShortcuts(match?.disableGlobalShortcuts);
+
   if (!element) {
     // error UI here probably
     return null;
@@ -677,6 +694,7 @@ function CurrentRoute(props: { pathname: string }) {
   const direction = isBack
     ? directionMap[previousDirection as Direction]
     : currentDirection;
+
   return (
     <AnimatePresence key={props.pathname} mode="popLayout">
       {React.cloneElement(element, {
@@ -694,4 +712,25 @@ const directionMap = {
   left: 'right',
   down: 'up',
   base: 'base',
+};
+
+const useGlobalShortcuts = (disable?: boolean) => {
+  const { sortedAccounts } = useAccounts();
+  const { setCurrentAddress } = useCurrentAddressStore();
+  useKeyboardShortcut({
+    handler: (e: KeyboardEvent) => {
+      const activeElement = document.activeElement;
+      const tagName = activeElement?.tagName;
+      if (tagName !== 'INPUT') {
+        const regex = /^[1-9]$/;
+        if (regex.test(e.key)) {
+          const accountIndex = parseInt(e.key, 10) - 1;
+          if (sortedAccounts[accountIndex]) {
+            setCurrentAddress(sortedAccounts[accountIndex]?.address);
+          }
+        }
+      }
+    },
+    condition: () => !disable,
+  });
 };
